Add files via upload
This is a diagram of two web servers I created in the Azure Resource Group. The purpose of this diagram is to demonstrate the best security practices in order to protect against DNS spoofing, malware injection, and the like using the network infrastructure of cloud services, a load balancer, and two barriers of ssh logins.